pyproject.toml
[project]
name = "acs-lung-cardiac-eda"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = ["polars", "ruff", "python-dateutil"]
